Upgrade cf cli to v7
Upgrading to version 7 of the CF cli as v6 is in the process of being
deprecated. According to
https://github.com/cloudfoundry/cli/wiki/Versioning-and-Support-Policy,
"All new features, enhancements, and fixes will be made on the v7 CLI.
The v6 CLI will only be updated to resolve the most severe defects
and/or security issues."

The installation method for installing the cli has also been changed
from Homebrew to using the compressed binary for speed purposes in the
github actions workflow.