Pin Postgres version in test workflow
With the latest (14.0) release of Postgres, the test workflow is failing
due to pg_dump server version mismatch errors. As the SORN project is
using an AWS RDS instance that runs on Postgres version 12.x the ci/cd
pipeline should never be using a different version. Pinning the version
in the test workflow should allow for much more stable and accurate
testing.