Add test success condition on deployments
Currently the deploy workflow will run on every completed run of the
test workflow which is dangerous at best. If the test workflow fails,
the deployment workflow should never be kicked off. This logic error can
be fixed by adding a conditional to the deploy action that checks the
success status of the triggering workflow_run. See
https://docs.github.com/en/actions/learn-github-actions/events-that-trigger-workflows#workflow_run