Return bad request status for null byte in params
A null byte sent in any of the parameters is unexpected and when
searching will cause some issues if passed through. To prevent that, this
checks all params of any request and sends back a 'bad request' status if a
null byte is found by raising the error and letting Rails do its thing.

Pattern based off of http://joshfrankel.me/blog/don-t-let-the-null-bytes-bite/