Fix agency filtering
An attribute (xml) we need for displaying the SORNS was accidentally
left out of the list of attributes the agency filtering scope returns.
To fix and avoid this in the future, we now select *all* database
attributes when using the scope.

Also updates a test to make sure code path it tests is triggered;
parameter was incorrectly named before, making it pass incorrectly.

Signed-off-by: Natasha Ibrahim <natasha.ibrahim@gsa.gov>