Prevent routing of requests with arbitrary formats
By adding "format: false" to the route configuration for static pages,
the application will no longer respond to requests of the form
"sorn.gov/help.abcd" or "sorn.gov/help.bac", etc. The only valid request
to a page such as the /help page is now /help with no extensions and/or
dot suffix.